check token format
